deepface-api/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
deepface>=0.0.89
python-multipart>=0.0.6
pillow>=10.2.0
numpy>=1.26.0
tf-keras>=2.15.0
